requirements.txt
langchain
langchain-community
slack-bolt
python-dotenv
requests
